# docs/swagger.yaml
definitions:
  account.ContactMailBinder:
    properties:
      email:
        type: string
      message:
        type: string
      name:
        type: string
      subject:
        type: string
    type: object
  account.LoginBinder:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  account.MailBinder:
    properties:
      email:
        type: string
    type: object
  account.ModifyBinder:
    properties:
      email:
        type: string
      is_superuser:
        type: string
      password:
        type: string
      phone:
        type: string
      username:
        type: string
    type: object
  account.ModifyInformationBinder:
    properties:
      address:
        type: string
      age:
        type: string
      avatar:
        type: string
      biography:
        type: string
      brith_date:
        type: string
      description:
        type: string
      nickname:
        type: string
      posts:
        type: string
      sex:
        type: string
      style:
        type: string
    type: object
  account.RegisterBinder:
    properties:
      code:
        type: string
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
  response.Information:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
  response.SuccessLogin:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      token:
        type: string
    type: object
info:
  contact: {}
  description: 在这里你可以获取想要的照片，并对照片进行二次创作，快来开始使用吧！
  title: 图片摄影创作社交网站
  version: "1.0"
paths:
  /account/contact/mail:
    post:
      consumes:
      - application/json
      description: 用户发送联系邮件
      parameters:
      - description: 发送联系邮件参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/account.ContactMailBinder'
      produces:
      - application/json
      responses:
        "200":
          description: 发送联系邮件成功
          schema:
            $ref: '#/definitions/response.Information'
        "401":
          description: 发送联系邮件失败
          schema:
            $ref: '#/definitions/response.Information'
      security:
      - ApiKeyAuth: []
      summary: 发送联系邮件
      tags:
      - 发送联系邮件
  /account/delete:
    get:
      consumes:
      - application/json
      description: 用于删除账号
      parameters:
      - description: 删除账号参数
        in: query
        name: username
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除账号成功
          schema:
            $ref: '#/definitions/response.Information'
        "401":
          description: 删除账号失败
          schema:
            $ref: '#/definitions/response.Information'
        "402":
          description: 请输入账号
          schema:
            $ref: '#/definitions/response.Information'
      security:
      - ApiKeyAuth: []
      summary: 删除账号
      tags:
      - 删除账号
  /account/login:
    post:
      consumes:
      - application/json
      description: 用于用户登录
      parameters:
      - description: 登录参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/account.LoginBinder'
      produces:
      - application/json
      responses:
        "200":
          description: 登陆成功
          schema:
            $ref: '#/definitions/response.SuccessLogin'
        "401":
          description: 账号或者密码错误
          schema:
            $ref: '#/definitions/response.Information'
        "402":
          description: 请输入账号或者密码
          schema:
            $ref: '#/definitions/response.Information'
        "403":
          description: 该用户未注册
          schema:
            $ref: '#/definitions/response.Information'
        "404":
          description: 更新时间失败
          schema:
            $ref: '#/definitions/response.Information'
      security:
      - ApiKeyAuth: []
      summary: 用户登录
      tags:
      - 登录
  /account/logout:
    get:
      consumes:
      - application/json
      description: 用于登录用户注销
      parameters:
      - description: 注销参数
        in: query
        name: email
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 注销成功
          schema:
            $ref: '#/definitions/response.Information'
        "401":
          description: 注销失败
          schema:
            $ref: '#/definitions/response.Information'
        "402":
          description: 您还未登录
          schema:
            $ref: '#/definitions/response.Information'
        "403":
          description: 请输入邮箱账号
          schema:
            $ref: '#/definitions/response.Information'
      security:
      - ApiKeyAuth: []
      summary: 用户注销
      tags:
      - 注销
  /account/mail:
    post:
      consumes:
      - application/json
      description: 新用户发送验证码用于注册账号
      parameters:
      - description: 发送邮件参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/account.MailBinder'
      produces:
      - application/json
      responses:
        "200":
          description: 发送验证码成功
          schema:
            $ref: '#/definitions/response.Information'
        "401":
          description: 邮件发送失败
          schema:
            $ref: '#/definitions/response.Information'
        "402":
          description: 请输入邮箱或者密码
          schema:
            $ref: '#/definitions/response.Information'
      security:
      - ApiKeyAuth: []
      summary: 发送验证码邮件
      tags:
      - 验证码
  /account/modify:
    post:
      consumes:
      - application/json
      description: 用于修改用户账号信息
      parameters:
      - description: 修改用户账号信息参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/account.ModifyBinder'
      produces:
      - application/json
      responses:
        "200":
          description: 修改信息成功
          schema:
            $ref: '#/definitions/response.Information'
        "401":
          description: 修改信息失败
          schema:
            $ref: '#/definitions/response.Information'
        "402":
          description: 请输入信息
          schema:
            $ref: '#/definitions/response.Information'
      security:
      - ApiKeyAuth: []
      summary: 修改用户账号信息
      tags:
      - 修改用户账号信息
  /account/modify/information:
    post:
      consumes:
      - application/json
      description: 用于修改用户详细信息
      parameters:
      - description: 修改用户详细信息参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/account.ModifyInformationBinder'
      produces:
      - application/json
      responses:
        "200":
          description: 修改信息成功
          schema:
            $ref: '#/definitions/response.Information'
        "401":
          description: 修改信息失败
          schema:
            $ref: '#/definitions/response.Information'
        "402":
          description: 请输入信息
          schema:
            $ref: '#/definitions/response.Information'
      security:
      - ApiKeyAuth: []
      summary: 修改用户详细信息
      tags:
      - 修改用户详细信息
  /account/register:
    post:
      consumes:
      - application/json
      description: 用于新用户注册账号使用
      parameters:
      - description: 注册参数
        in: body
        name: object
        schema:
          $ref: '#/definitions/account.RegisterBinder'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            $ref: '#/definitions/response.Information'
        "401":
          description: 账号已注册
          schema:
            $ref: '#/definitions/response.Information'
        "402":
          description: 验证码错误
          schema:
            $ref: '#/definitions/response.Information'
        "403":
          description: 请输入完整的信息
          schema:
            $ref: '#/definitions/response.Information'
      security:
      - ApiKeyAuth: []
      summary: 新用户注册
      tags:
      - 注册
  /author:
    get:
      consumes:
      - application/json
      description: 用于所有摄影师信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取所有摄影师信息成功
          schema:
            $ref: '#/definitions/response.Information'
        "401":
          description: 获取所有摄影师信息失败
          schema:
            $ref: '#/definitions/response.Information'
      security:
      - ApiKeyAuth: []
      summary: 显示所有摄影师信息
      tags:
      - 显示所有摄影师信息
  /gallery:
    get:
      consumes:
      - application/json
      description: 用于显示主题图片信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取主题图片信息成功
          schema:
            $ref: '#/definitions/response.Information'
        "401":
          description: 获取主题图片信息失败
          schema:
            $ref: '#/definitions/response.Information'
      security:
      - ApiKeyAuth: []
      summary: 显示主题图片信息
      tags:
      - 显示图片详细信息
  /gallery-single:
    get:
      consumes:
      - application/json
      description: 用于显示图片详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取图片详细信息成功
          schema:
            $ref: '#/definitions/response.Information'
        "401":
          description: 获取图片详细信息失败
          schema:
            $ref: '#/definitions/response.Information'
      security:
      - ApiKeyAuth: []
      summary: 显示图片详细信息
      tags:
      - 显示图片详细信息
  /image/index:
    get:
      consumes:
      - application/json
      description: 用于显示主页图片
      produces:
      - application/json
      responses:
        "200":
          description: 获取主页图片成功
          schema:
            $ref: '#/definitions/response.Information'
        "401":
          description: 获取主页图片失败
          schema:
            $ref: '#/definitions/response.Information'
      security:
      - ApiKeyAuth: []
      summary: 显示主页图片
      tags:
      - 显示主页图片
  /mine:
    get:
      consumes:
      - application/json
      description: 用于摄影师详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取摄影师详细信息成功
          schema:
            $ref: '#/definitions/response.Information'
        "401":
          description: 获取摄影师详细信息失败
          schema:
            $ref: '#/definitions/response.Information'
      security:
      - ApiKeyAuth: []
      summary: 显示摄影师详细信息
      tags:
      - 显示摄影师详细信息
swagger: "2.0"